drain: refactor batch_future into its own file
aka What If structs.go Wasn't So Big?
